r009: bug fix for queue and subgraph labeling; added a non-covcut single-tig circle cut (which now also tolerates multiple outward arcs as long as they are on the same end) after complex bubble popping.

+++ Overlaps.h
@@ -115,7 +115,7 @@ typedef struct {
 	uint64_t *a; // list of reads
 	char *s; // unitig sequence is not null
 	uint8_t c;  // hamt: like seq's .c
-	uint8_t subg_label;  // hamt: subgraph ID
+	int subg_label;  // hamt: subgraph ID
 } ma_utg_t;
